refactor: move InstalledPackae & FetchedPackage types to types mod
